refactor(Shape): 카드 모양을 Enum으로 대체
