+++ README.md
@@ -7,7 +7,8 @@
 - [x] 카드의 숫자 계산은 카드 문양이 아닌 카드 숫자의 합으로 한다.
 - [x] King, Queen, Jack은 각각 10으로 계산한다.
 - [x] 예외로 Ace는 카드의 합이 21을 초과하기 전까지는 11로 계산하고, 그 외에는 1로 계산한다.
-- [ ] 게임을 시작하면 플레이어는 두 장의 카드를 지급 받는다.
+- [-] 카드 생성기는 중복되지 않는 임의의 카드를 생성한다.
+- [x] 게임을 시작하면 플레이어는 두 장의 카드를 지급 받는다.
 - [ ] 21을 넘지 않을 경우 원한다면 얼마든지 카드를 계속 뽑을 수 있다.
 - [ ] 딜러는 처음에 받은 2장의 합계가 16이하이면 반드시 1장의 카드를 추가로 받아야 한다.
 - [ ] 17점 이상이면 카드를 추가로 받지 않는다.
