+++ src/main/java/model/Shape.java
@@ -1,0 +1,8 @@
+package model;
+
+public enum Shape {
+    SPADE,
+    HEART,
+    DIAMOND,
+    CLOVER
+}
